+++ src/main/java/bit/project/lawbot/service/CommentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface CommentService {
 	public List<CommentDTO> selectCommentList(CommentDTO dto);
-	public int insertComment(CommentDTO dto);
+	public long insertComment(CommentDTO dto);
 	public CommentDTO selectComment(CommentDTO dto);
 	public CommentDTO deleteComment(CommentDTO dto);
 	public CommentDTO updateComment(CommentDTO dto);
